requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
fastapi==0.104.0
uvicorn==0.24.0
streamlit==1.28.0
plotly==5.17.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
joblib==1.3.2
imbalanced-learn==0.11.0
pytest==7.4.2
httpx==0.25.0
kaggle==1.5.16
